Comments - Defect Number: 2984; Defect RecordID: 2980; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: New Time Entry Reporting.; Defect Description: From: Jason Hong-Turney [mailto:jasonht@logic2020.com] Sent: Tuesday, February 28, 2012 12:22 PM To: MadhanA (madhana@ggktech.com) Cc: logic2020@ggktech.com; Matt Reilly Subject: Time Entry reporting changes
Hi Madhan 

Ive redesigned the Time Entry reporting to reflect the changes were making in Time Entry. My goal is to make the report output simpler, especially for Excel exports. I also tried to add some information and graphs that will make it the time entries were capturing more relevant to our business, and that will be easier to get summary/high-level data from. Ive included placeholders for PDF exports, but these can be worked on later. The Excel reports will be more important initially.

Begin by looking at the Time Period Summary image (it uses the correct top selection bar text, for example) before looking at the other two images. The Time Period Summary and Project Summary are somewhat similar in design, while the Person Detail reporting is more specific to the information it contains.

Please look at these and let me know what questions you have and/or what I need to clarify. I dont expect the team to work on these changes until the Time Entry changes I sent earlier and the Time Off functionality changes are completed. In the meantime, it should give you and I a chance to get this as clear as we can.

Jason Hong-Turney // Logic20/20 // jasonht@logic2020.com
t.206-576-0472 // c. 425-390-4664

From: MadhanA [mailto:madhana@ggktech.com]
Sent: Thursday, March 01, 2012 9:33 AM
To: Jason Hong-Turney
Cc: logic2020@ggktech.com; Matt Reilly
Subject: RE: Time Entry reporting changes

Hi Jason,

Can you please provide your answers to the below questions.

Time Period Summary Report-> B vs NB Graph %:

1. we assumed that the Exclude Admin group means the time entries entered in Administrative section should be excluded. Please let us know if we are wrong.
[Jason]  What I intended was to exclude the time entries from people in the Administrative practice (the internal employees here).

Project Summary Report:

1. As per existing functionality, person may be assigned to a milestone/Project with multiple roles. In that scenario, How do you want us to handle(Like showing all roles in project Role column as comma separated list or Showing only the highest role in the ladder?
[Jason] Show only the highest role for now. I plan to change this in the future, to better align the Team Structure options with the role options.
2. There is no date range selector in this Report, so we assumed that the date range should be the project start date and end date. Please let us know if we are wrong.
[Jason] Theres no date range because the report will just look at the Project overall, so I dont want any date filtering options at this point. Were going to be moving away from the concept of milestones and people being linked at that level. Resources will simply be linked to the Project. Were still working on these parts and thinking through how calculations, etc, are affected.
--
Thanks & Regards,
Madhan Annabemoju / GGK Technologies

From: Jason Hong-Turney [mailto:jasonht@logic2020.com]
Sent: Friday, March 02, 2012 6:53 AM
To: MadhanA
Cc: logic2020@ggktech.com
Subject: RE: Time Entry reporting changes

Hi Madhan  see answers below!

Jason Hong-Turney // Logic20/20 // jasonht@logic2020.com
t.206-576-0472 // c. 425-390-4664

Hi Jason,

Can you please provide your answers to the below questions.

Questions:
Project Summary Report:

1) As per the new time entry functionality user is able to enter time for a project outside its start date and end date. In this case do we need to show those time entries?
[Jason] Yes, any time that is applied towards a Project should be shown.
2) In the same way the person who

+++ Logic2020, Inc./Practice Management/DotNet/PracticeManagement/ClientDAL/ReportDAL.cs
@@ -261,8 +261,8 @@ namespace DataAccess
                         GroupedHours GH = new GroupedHours();
                         GH.StartDate = reader.GetDateTime(startDateIndex);
                         GH.SetEnddate(reader.GetString(groupByCerteriaIndex));
-                        GH.BillabileTotal = reader.GetInt32(billableHoursIndex);
-                        GH.NonBillableTotal = reader.GetInt32(nonBillableHoursIndex);
+                        GH.BillabileTotal = reader.GetDouble(billableHoursIndex);
+                        GH.NonBillableTotal = reader.GetDouble(nonBillableHoursIndex);
 
                         PLGH.GroupedHoursList = new List<GroupedHours>();
                         PLGH.GroupedHoursList.Add(GH);
@@ -274,8 +274,8 @@ namespace DataAccess
                         GroupedHours GH = new GroupedHours();
                         GH.StartDate = reader.GetDateTime(startDateIndex);
                         GH.SetEnddate(reader.GetString(groupByCerteriaIndex));
-                        GH.BillabileTotal = reader.GetInt32(billableHoursIndex);
-                        GH.NonBillableTotal = reader.GetInt32(nonBillableHoursIndex);
+                        GH.BillabileTotal = reader.GetDouble(billableHoursIndex);
+                        GH.NonBillableTotal = reader.GetDouble(nonBillableHoursIndex);
 
                         PLGH.GroupedHoursList.Add(GH);
                     }
@@ -341,8 +341,8 @@ namespace DataAccess
                         GroupedHours GH = new GroupedHours();
                         GH.StartDate = reader.GetDateTime(startDateIndex);
                         GH.SetEnddate(reader.GetString(groupByCerteriaIndex));
-                        GH.BillabileTotal = reader.GetInt32(billableHoursIndex);
-                        GH.NonBillableTotal = reader.GetInt32(nonBillableHoursIndex);
+                        GH.BillabileTotal = reader.GetDouble(billableHoursIndex);
+                        GH.NonBillableTotal = reader.GetDouble(nonBillableHoursIndex);
 
                         plgh.GroupedHoursList = new List<GroupedHours>();
                         plgh.GroupedHoursList.Add(GH);
@@ -354,8 +354,8 @@ namespace DataAccess
                         GroupedHours gh = new GroupedHours();
                         gh.StartDate = reader.GetDateTime(startDateIndex);
                         gh.SetEnddate(reader.GetString(groupByCerteriaIndex));
-                        gh.BillabileTotal = reader.GetInt32(billableHoursIndex);
-                        gh.NonBillableTotal = reader.GetInt32(nonBillableHoursIndex);
+                        gh.BillabileTotal = reader.GetDouble(billableHoursIndex);
+                        gh.NonBillableTotal = reader.GetDouble(nonBillableHoursIndex);
 
                         plgh.GroupedHoursList.Add(gh);
                     }
